requirements.txt
requests~=2.32.3
python-telegram-bot~=22.1
psycopg2-binary~=2.9.10
pymorphy3~=2.0.3